lazily initialize the registry at the first use + fix mypy observability error

+++ shared/observability/src/airflow_shared/observability/metrics/base_stats_logger.py
@@ -112,7 +112,7 @@ class NoStatsLogger:
     def gauge(
         cls,
         stat: str,
-        value: int,
+        value: float,
         rate: int = 1,
         delta: bool = False,
         *,
